+++ YandexMusicResolver/Requests/YandexCustomRequest.cs
@@ -1,6 +1,9 @@
-﻿namespace YandexMusicResolver.Requests {
+﻿using YandexMusicResolver.Config;
+
+namespace YandexMusicResolver.Requests {
     internal class YandexCustomRequest : YandexRequest {
-        public YandexCustomRequest(string? token = null) : base(token) { }
+        public YandexCustomRequest(IYandexProxyHolder? proxyHolder, IYandexTokenHolder? tokenHolder) : base(proxyHolder, tokenHolder) { }
+        public YandexCustomRequest(IYandexProxyTokenHolder? config = null) : base(config) { }
 
         public YandexCustomRequest Create(string url) {
             FormRequest(url);