+++ RELEASE-NOTES.md
@@ -1,3 +1,10 @@
+# v2.0.0
+- Add new config system
+- Query parser in yandex searches now separated
+- Fixes for a main resolver, playlist tracks resolving, prefixes resolving
+- Add Github Actions workflows
+- Add unit tests
+
 # v1.1.0
 - Add authentications via YandexMusicAuth class
 - New internal request structure